Fix: added the code_preview function in another file
